requirements.txt

Django==1.8
MySQL-python==1.2.5
djangorestframework==2.3.14
djutil==0.2
requests
simplejson
django-bootstrap-form
pyotp